feat: Add kanji candidate support with IPADIC dictionary
- Integrate IPADIC dictionary for kanji conversion
- Refactor web app to support kanji candidates
- Add lattice-based path finding for kanji conversion
- Update data types for better memory usage (u32 for IDs, i16 for costs)
- Increase WASM memory pages for dictionary support
- Restructure React app with routing and new components

+++ tests/dictionary.test.zig
@@ -51,7 +51,7 @@ fn createTestDictionary(allocator: mem.Allocator) !Dictionary {
     var ltrie = try bldr.build();
     errdefer ltrie.deinit();
 
-    const costs = std.ArrayList(isize).init(allocator);
+    const costs = std.ArrayList(i16).init(allocator);
 
     return .{
         .trie = ltrie,
